feat: Cache logger command existence for performance improvements
This commit introduces a performance optimization by caching the existence of the `Invoke-OSDCloudLogger` command in the `$script:LoggerExists` variable at the module level. This eliminates redundant calls to `Get-Command` within functions, resulting in improved execution speed, especially in scenarios where logging is frequently used. The cached value is then used in `if` statements to determine whether to call the logger. A verbose message is also added to confirm the state of the cached variable.